Allow running superflore with uncached rosdistro
This commits adds a new command line flag to force superflore (via the
rosdistro python package) to regenerate the distribution information.

This is useful when running superflore with $ROSDISTRO_INDEX_URL set to
get deterministic codegen. The rosdistro caches are updated to latest
and not versioned, see https://github.com/ros/rosdistro/issues/47469.